fix(plugins): preserve lightweight runtime scope imports
